requirements.txt
dlib >= 19.9.0
numpy >= 1.13.1
scipy >= 0.18.0
Pillow
opencv-python-headless
dlib-bin
streamlit